third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-element-repeat-get-set.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<link href="resources/grid.css" rel="stylesheet">
<style>
#singleSingleTrackRepeat {
    grid-definition-rows: repeat(1, 18px);
    grid-definition-columns: repeat(1, 15%);
}

#twoSingleTrackRepeat {
    grid-definition-rows: repeat(2, auto);
    grid-definition-columns: repeat(2, minmax(15px, 50%));
}

#twoDoubleTrackRepeat {
    grid-definition-rows: repeat(2, minmax(5px, 10px)  auto);
    grid-definition-columns: repeat(2, auto minmax(100px, 120px));
}

#twoDoubleTrackWithNamedGridLineRepeat {
    grid-definition-rows: repeat(2, 10px 'start' auto 'end');
    grid-definition-columns: repeat(2, auto 'middle' 250px 'end');
}

#twoDoubleTrackWithTrailingNamedGridLineRepeat {
    grid-definition-rows: repeat(2, 'before' 10px);
    grid-definition-columns: repeat(2, 'before' auto);
}

#trailingNamedGridLineRepeat {
    grid-definition-rows: repeat(1, 10px) 'end';
    grid-definition-columns: repeat(1, 250px) 'end';
}

#leadingNamedGridLineRepeat {
    grid-definition-rows: 'start' repeat(2, 10px);
    grid-definition-columns: 'start' repeat(2, 250px);
}

#mixRepeatAfterNonRepeat {
    grid-definition-rows: auto repeat(2, 10px);
    grid-definition-columns: 'start' 140px repeat(2, 250px);
}

#mixNonRepeatAfterRepeat {
    grid-definition-rows: repeat(2, 10px) 'end' auto;
    grid-definition-columns: repeat(2, 250px) 15% 'last';
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div class="grid" id="singleSingleTrackRepeat"></div>
<div class="grid" id="twoSingleTrackRepeat"></div>
<div class="grid" id="twoDoubleTrackRepeat"></div>
<div class="grid" id="twoDoubleTrackWithNamedGridLineRepeat"></div>
<div class="grid" id="twoDoubleTrackWithTrailingNamedGridLineRepeat"></div>
<div class="grid" id="trailingNamedGridLineRepeat"></div>
<div class="grid" id="leadingNamedGridLineRepeat"></div>
<div class="grid" id="mixRepeatAfterNonRepeat"></div>
<div class="grid" id="mixNonRepeatAfterRepeat"></div>
<script>
    description('Test that setting and getting grid-definition-columns and grid-definition-rows with repeat() works as expected');

    debug("Test getting grid-definition-columns and grid-definition-rows set through CSS");
    var singleSingleTrackRepeat = document.getElementById("singleSingleTrackRepeat")
    shouldBeEqualToString("window.getComputedStyle(singleSingleTrackRepeat, '').getPropertyValue('grid-definition-rows')", "18px");
    shouldBeEqualToString("window.getComputedStyle(singleSingleTrackRepeat, '').getPropertyValue('grid-definition-columns')", "15%");

    var twoSingleTrackRepeat = document.getElementById("twoSingleTrackRepeat")
    shouldBeEqualToString("window.getComputedStyle(twoSingleTrackRepeat, '').getPropertyValue('grid-definition-rows')", "auto auto");
    shouldBeEqualToString("window.getComputedStyle(twoSingleTrackRepeat, '').getPropertyValue('grid-definition-columns')", "minmax(15px, 50%) minmax(15px, 50%)");

    var twoDoubleTrackRepeat = document.getElementById("twoDoubleTrackRepeat")
    shouldBeEqualToString("window.getComputedStyle(twoDoubleTrackRepeat, '').getPropertyValue('grid-definition-rows')", "minmax(5px, 10px) auto minmax(5px, 10px) auto");
    shouldBeEqualToString("window.getComputedStyle(twoDoubleTrackRepeat, '').getPropertyValue('grid-definition-columns')", "auto minmax(100px, 120px) auto minmax(100px, 120px)");

    var twoDoubleTrackWithNamedGridLineRepeat = document.getElementById("twoDoubleTrackWithNamedGridLineRepeat")
    shouldBeEqualToString("window.getComputedStyle(twoDoubleTrackWithNamedGridLineRepeat, '').getPropertyValue('grid-definition-rows')", "10px start auto end 10px start auto end");
    shouldBeEqualToString("window.getComputedStyle(twoDoubleTrackWithNamedGridLineRepeat, '').getPropertyValue('grid-definition-columns')", "auto middle 250px end auto middle 250px end");

    var twoDoubleTrackWithTrailingNamedGridLineRepeat = document.getElementById("twoDoubleTrackWithTrailingNamedGridLineRepeat")
    shouldBeEqualToString("window.getComputedStyle(twoDoubleTrackWithTrailingNamedGridLineRepeat, '').getPropertyValue('grid-definition-rows')", "before 10px before 10px");
    shouldBeEqualToString("window.getComputedStyle(twoDoubleTrackWithTrailingNamedGridLineRepeat, '').getPropertyValue('grid-definition-columns')", "before auto before auto");

    var trailingNamedGridLineRepeat = document.getElementById("trailingNamedGridLineRepeat")
    shouldBeEqualToString("window.getComputedStyle(trailingNamedGridLineRepeat, '').getPropertyValue('grid-definition-rows')", "10px end");
    shouldBeEqualToString("window.getComputedStyle(trailingNamedGridLineRepeat, '').getPropertyValue('grid-definition-columns')", "250px end");

    var leadingNamedGridLineRepeat = document.getElementById("leadingNamedGridLineRepeat")
    shouldBeEqualToString("window.getComputedStyle(leadingNamedGridLineRepeat, '').getPropertyValue('grid-definition-rows')", "start 10px 10px");
    shouldBeEqualToString("window.getComputedStyle(leadingNamedGridLineRepeat, '').getPropertyValue('grid-definition-columns')", "start 250px 250px");

    var mixRepeatAfterNonRepeat = document.getElementById("mixRepeatAfterNonRepeat")
    shouldBeEqualToString("window.getComputedStyle(mixRepeatAfterNonRepeat, '').getPropertyValue('grid-definition-rows')", "auto 10px 10px");
    shouldBeEqualToString("window.getComputedStyle(mixRepeatAfterNonRepeat, '').getPropertyValue('grid-definition-columns')", "start 140px 250px 250px");

    var mixNonRepeatAfterRepeat = document.getElementById("mixNonRepeatAfterRepeat")
    shouldBeEqualToString("window.getComputedStyle(mixNonRepeatAfterRepeat, '').getPropertyValue('grid-definition-rows')", "10px 10px end auto");
    shouldBeEqualToString("window.getComputedStyle(mixNonRepeatAfterRepeat, '').getPropertyValue('grid-definition-columns')", "250px 250px 15% last");

    debug("");
    debug("Test invalid repeat syntax.");
    function testInvalidSyntax(gridColumn) {
        element = document.createElement("div");
        document.body.appendChild(element);
        element.style.gridDefinitionColumns = gridColumn;
        shouldBeEqualToString("window.getComputedStyle(element, '').getPropertyValue('grid-definition-columns')", "none");
        document.body.removeChild(element);
    }
    testInvalidSyntax("repeat(");
    testInvalidSyntax("repeat()");
    testInvalidSyntax("repeat(3 / auto)");
    testInvalidSyntax("repeat(3 , ,)");
    testInvalidSyntax("repeat(0, 15px)");
    testInvalidSyntax("repeat(-1, auto)");
    // Nesting is no allowed.
    testInvalidSyntax("repeat(2, repeat(1, auto))");
</script>
</body>
</html>
